refactor: replace multiple Supabase clients with shared SupabaseService
