+++ src/app/services/supabase/game-supabase.service.ts
@@ -1,24 +1,17 @@
-import type { SupabaseClient } from '@supabase/supabase-js';
-import { createClient } from '@supabase/supabase-js';
-import { Injectable } from '@angular/core';
-import { environment } from '@/environments/environment.development';
+import { inject, Injectable } from '@angular/core';
 import { GAME_ID } from '@/app/constants/auth.constants';
 import type { GameModel, GameProjection } from '@/app/types/supabase-game.type';
 import { Chess } from 'chess.js';
 import type { GameResultType } from '@/app/services/game.service';
 import { clone } from '@/app/utilities/chess-piece';
 import type { MoveRecordType } from '@/app/store/states/game.state';
+import { SupabaseService } from '@/app/services/supabase/supabase.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class GameSupabaseService {
-  private readonly apiUrl = environment.apiUrl;
-  private readonly publishableKey = environment.publishableKey;
-  private readonly supabase: SupabaseClient = createClient(
-    this.apiUrl,
-    this.publishableKey,
-  );
+  private readonly supabase = inject(SupabaseService).client;
 
   public async createGame(
     playerId: string,
